ref: разделение на ari/ami
