+++ src/asterisk/ami/ami.constants.ts
@@ -1,3 +1,5 @@
+import { DNDApiStatus, DNDStatus, HintStatus } from './interfaces/ami.enum';
+
 export const DEFAULT_REOPEN_AMI_CLIENT = 5000;
 export const AMI_RECONECT = 'Переподключение к AMI ...';
 export const AMI_INCORRECT_LOGIN = 'Некорректный логин или пароль от AMI';
@@ -9,3 +11,13 @@ export const NOT_LOCAL_NUMBER = 4;
 export const DEFAULT_HANGUP_CDR_TIMEOUT = 1000;
 export const MIN_EXTERNAL_NUMBER_LENGTH = 10;
 export const DEFAULT_TIMEOUT_HANDLER = 2000;
+
+export const DND_API_TO_DND_STATUS: { [code in DNDApiStatus]: DNDStatus } = {
+  [DNDApiStatus.on]: DNDStatus.on,
+  [DNDApiStatus.off]: DNDStatus.off,
+};
+
+export const DND_API_TO_HINT_STATUS: { [code in DNDApiStatus]: HintStatus } = {
+  [DNDApiStatus.on]: HintStatus.on,
+  [DNDApiStatus.off]: HintStatus.off,
+};
