+++ src/asterisk/ami/services/action-service.ts
@@ -2,17 +2,12 @@ import { IDnd } from '@app/asterisk-api/interfaces/asterisk-api.interfaces';
 import { Injectable } from '@nestjs/common';
 import * as namiLib from 'nami';
 import { AsteriskAmi } from '../ami';
-import {
-  AsteriskDNDStatusResponse,
-  AsteriskStatusResponse,
-  dndStatusMap,
-  EventsStatus,
-  hintStatusMap,
-  SetDNDStatusResult,
-} from '../../interfaces/asterisk.interfaces';
-import { ChannelType, DbFamilyType, statusHint } from '../../interfaces/asterisk.enum';
 import { LogService } from '@app/log/log.service';
 import { AMI_OUTBOUND_CALL } from '@app/asterisk/ari/ari.constants';
+import { AsteriskDNDStatusResponse, AsteriskStatusResponse, EventsStatus, SetDNDStatusResult } from '../interfaces/ami.interfaces';
+import { DND_API_TO_DND_STATUS, DND_API_TO_HINT_STATUS } from '../ami.constants';
+import { ChannelType } from '@app/asterisk/ari/interfaces/ari.enum';
+import { DbFamilyType, HintStatus } from '../interfaces/ami.enum';
 
 @Injectable()
 export class AmiActionService {
@@ -39,7 +34,7 @@ export class AmiActionService {
     }
   }
 
-  private async setHintStatus(extension: string, hint: statusHint): Promise<void> {
+  private async setHintStatus(extension: string, hint: HintStatus): Promise<void> {
     try {
       const action = new namiLib.Actions.Command();
       action.Command = `devstate change Custom:DND${extension} ${hint}`;
@@ -62,7 +57,7 @@ export class AmiActionService {
           }
           const resultSend: AsteriskStatusResponse = await this.dndPut(sip_id, data.dnd_status);
           this.log.info(resultSend, AmiActionService.name);
-          const hint = hintStatusMap[data.dnd_status];
+          const hint = DND_API_TO_HINT_STATUS[data.dnd_status];
 
           if (resultSend.response == 'Success') {
             extensionStatusList[sip_id] = { status: 'success' };
@@ -84,7 +79,7 @@ export class AmiActionService {
     const action = new namiLib.Actions.DbPut();
     action.Family = DbFamilyType.DND;
     action.Key = sipId;
-    action.Val = dndStatusMap[dndStatus];
+    action.Val = DND_API_TO_DND_STATUS[dndStatus];
     return await this.ami.amiClientSend(action);
   }
